refactor(interview+case): refactor to use custom react hooks for readability
